HyperMetro-5 Split CommandParser into two classes
Move methods not related to reading and parsing the commands to a new class
called Controller.

CommandParser.java
- Move the start(), getRoute() and printRoute() methods to the new Controller
  class.
- Update the constructor and remove unneeded fields.

Main.java
- Update readFile to use the new class.

Controller.java
- New class with fields and methods moved out of CommandParser.java

HyperMetroTests.java
- Update the tests to use the new classes.
